Common: add getLogger method to PluginUtils, add a bunch of enum/keyed helper methods to ConfigUtils Config: add compact list building, automatically detect collection on generic append I18n: remove useless methods Configurables: extract some common attributes to attribute/common package Make AttributeFactory accept generic of attribute Add paramaterless attribute map constructor Remove deprecated configurables Add more verbose precondition to some of the attributes

+++ configurables/src/main/java/me/wyne/wutils/config/configurables/animation/AnimationTimings.java
@@ -1,11 +1,11 @@
 package me.wyne.wutils.config.configurables.animation;
 
 public class AnimationTimings {
-    public int delay;
-    public int period;
-    public int duration;
+    public long delay;
+    public long period;
+    public long duration;
 
-    public AnimationTimings(int delay, int period, int duration) {
+    public AnimationTimings(long delay, long period, long duration) {
         this.delay = delay;
         this.period = period;
         this.duration = duration;
